tests: Feature and unit tests

+++ api/app/Transformers/OpenWeatherTransformer.php
@@ -13,17 +13,17 @@ class OpenWeatherTransformer extends AbstractWeatherTransformer
 
         return [
             'location' => implode(", ", array_filter([$city, $country], fn ($item) => !empty($item))),
-            'current_temp' => round(optional($this->weather->current['main'])['temp']) . "°C",
+            'current_temp' => round((float)optional($this->weather->current['main'])['temp']) . "°C",
         ];
     }
 
     public function getDetails(): array
     {
         return [
-            'current_temp' => round(optional($this->weather->current['main'])['temp']) . "°C",
-            'min_temp' => round(optional($this->weather->current['main'])['temp_min']) . "°C",
-            'max_temp' => round(optional($this->weather->current['main'])['temp_max']) . "°C",
-            'feels_like' => round(optional($this->weather->current['main'])['feels_like']) . "°C",
+            'current_temp' => round((float)optional($this->weather->current['main'])['temp']) . "°C",
+            'min_temp' => round((float)optional($this->weather->current['main'])['temp_min']) . "°C",
+            'max_temp' => round((float)optional($this->weather->current['main'])['temp_max']) . "°C",
+            'feels_like' => round((float)optional($this->weather->current['main'])['feels_like']) . "°C",
             'pressure' => optional($this->weather->current['main'])['pressure'],
             'humidity' => optional($this->weather->current['main'])['humidity'],
             'status' => optional($this->weather->current['weather'][0])['main'],
